requirements.txt
Django==1.10.1
django-naomi==0.8
psycopg2==2.6.2
Unipath==1.1
dj-database-url==0.4.1
django-cors-headers==1.2.2
django-extensions==1.7.4
whitenoise==3.2.1
dj-static==0.0.6
django-suit==0.2.21
rows==0.3.0
python-dotenv==0.6.0
django-nested-admin==3.0.10
gunicorn==19.6.0
django-naomi==0.8
djangorestframework==3.4.7
django-rq==0.9.2
PyYAML==3.12
